+++ x64barebones/Userland/SampleCodeModule/Test/test_util.c
@@ -66,7 +66,7 @@ void endless_loop() {
 }
 
 void endless_loop_print(uint64_t wait) {
-  int64_t pid = get_pid();
+  int16_t pid = get_pid();
 
   while (1) {
     printf("%d ", pid);
